requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.7
langchain-text-splitters>=0.0.1
google-generativeai>=0.3.0
pdf2image>=1.16.3
pillow>=10.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tqdm>=4.65.0
numpy>=1.24.0